Refactor: Change fee type

+++ src/main/kotlin/model/Receipt.kt
@@ -6,7 +6,7 @@ import java.time.LocalDateTime
 data class Receipt(
     val receiptNumber: Int,
     val parkedDuration: Duration,
-    val fee: Int,
+    val fee: Long,
     val vehicleType: VehicleType,
     val entryTime: LocalDateTime,
     val exitTime: LocalDateTime
